module: Avoid infinite looping The Right Way(tm)
The "last_host" business removed in this commit was there because
with the previous code the root host of a chain would otherwise
return itself as a parent, which was an effect of us not properly
initializing the one entry in the "tcm_parent" struct to NULL
upon each call.

Since we never had an opportunity to overwrite the stack space
where it was stored, it meant a host with no parents would
return the last parent returned (even if it came from a
different parent chain).

With this patch, we initialize the struct properly and thus
can get rid of the monkey business with 'last_host' inside the
calling loop.

While at it, we also fix some compiler warnings in both
module.c and qh-connect.c

Signed-off-by: Andreas Ericsson <ageric79@gmail.com>